.aitk/requirements/requirements-QNN.txt
# huggingface-hub[hf_xet]==0.34.4
huggingface-hub[hf_xet]==0.34.4
filelock==3.18.0
fsspec==2024.12.0
packaging==24.2
pyyaml==6.0.2
requests==2.32.3
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
tqdm==4.67.1
colorama==0.4.6
typing-extensions==4.15.0
# onnx==1.17.0
onnx==1.17.0
numpy==2.2.4
protobuf==3.20.3
# olive-ai@git+https://github.com/microsoft/Olive.git@19abbd99463db9f608e3124237c1ecc74ac6e92e#egg=olive-ai
olive-ai@git+https://github.com/microsoft/Olive.git@19abbd99463db9f608e3124237c1ecc74ac6e92e#egg=olive-ai
numpy==2.2.4
onnx==1.17.0
onnx-ir==0.1.5
ml-dtypes==0.5.3
onnxscript==0.3.2
optuna==4.2.1
alembic==1.16.5
mako==1.3.10
markupsafe==3.0.2
sqlalchemy==2.0.43
greenlet==3.2.4
colorlog==6.9.0
pandas==2.2.3
python-dateutil==2.9.0.post0
six==1.17.0
pytz==2025.2
tzdata==2025.2
pydantic==2.11.1
annotated-types==0.7.0
pydantic-core==2.33.0
typing-inspection==0.4.1
torch==2.7.0
jinja2==3.1.6
networkx==3.5
setuptools==80.9.0
sympy==1.14.0
mpmath==1.3.0
torchmetrics==1.7.0
lightning-utilities==0.15.2
transformers==4.50.3
regex==2025.9.1
safetensors==0.6.2
tokenizers==0.21.4
# tabulate==0.9.0
tabulate==0.9.0
# datasets==3.5.0
datasets==3.5.0
aiohttp==3.11.16
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.7.0
attrs==25.3.0
multidict==6.6.4
propcache==0.3.2
yarl==1.20.1
dill==0.3.8
multiprocess==0.70.16
pyarrow==19.0.1
xxhash==3.5.0
# torchvision==0.22.0
torchvision==0.22.0
pillow==11.2.1
# onnxruntime-qnn==1.21.1
onnxruntime-qnn==1.21.1
coloredlogs==15.0.1
humanfriendly==10.0
pyreadline3==3.5.4
flatbuffers==25.2.10
# uvpip:install onnxruntime-genai==0.7.0 --no-deps;post
# not in requires
hf-xet==1.1.9
